Profile UI updated
