+++ src/app/icons/PencilIcon.tsx
@@ -1,0 +1,23 @@
+const PencilIcon = () => {
+  return (
+    <>
+      <svg
+        width="30"
+        height="30"
+        viewBox="0 0 30 30"
+        fill="none"
+        xmlns="http://www.w3.org/2000/svg"
+      >
+        <g id="Pencil">
+          <path
+            id="Vector"
+            d="M26.9695 8.26766L21.7336 3.03055C21.516 2.81282 21.2575 2.64009 20.9731 2.52225C20.6887 2.4044 20.3839 2.34375 20.076 2.34375C19.7681 2.34375 19.4633 2.4044 19.1789 2.52225C18.8944 2.64009 18.636 2.81282 18.4184 3.03055L3.96798 17.4821C3.74954 17.6991 3.57636 17.9573 3.45848 18.2418C3.3406 18.5262 3.28036 18.8312 3.28126 19.1391V24.3763C3.28126 24.9979 3.52819 25.594 3.96773 26.0335C4.40727 26.4731 5.00341 26.72 5.62501 26.72H10.8621C11.17 26.7208 11.475 26.6606 11.7594 26.5427C12.0439 26.4248 12.3021 26.2517 12.5192 26.0333L26.9695 11.5817C27.4089 11.1422 27.6558 10.5462 27.6558 9.92469C27.6558 9.30322 27.4089 8.70718 26.9695 8.26766ZM10.8984 21.095L19.2188 12.7747L20.5078 14.0638L12.1875 22.3841L10.8984 21.095ZM8.90626 19.1028L7.6172 17.8138L15.9375 9.49344L17.2266 10.7825L8.90626 19.1028ZM6.09376 20.2747L7.91134 22.0923L9.72657 23.9075H6.09376V20.2747ZM22.5 12.0716L17.9297 7.50126L20.0789 5.35204L24.6492 9.92235L22.5 12.0716Z"
+            fill="#2B3674"
+          />
+        </g>
+      </svg>
+    </>
+  );
+};
+
+export default PencilIcon;
